dispatcher: move creating upstream connections to a separate thread
Because we need to hold the lock over connections, we cannot extend the
connections array (addconnection) from the same thread as the one that's
walking over the connections array.  So instead, signal a connector
thread that deals with this, such that all can wait while the
connections array is realloced.